+++ temba/msgs/models.py
@@ -154,7 +154,6 @@ class Media(models.Model):
             uuid=uuid,
             org=org,
             url=public_file_storage.url(path),
-            name=filename,
             content_type=content_type,
             path=path,
             size=size,
@@ -479,12 +478,14 @@ class Msg(models.Model):
     )
 
     FAILED_SUSPENDED = "S"
+    FAILED_CONTACT = "C"
     FAILED_LOOPING = "L"
     FAILED_ERROR_LIMIT = "E"
     FAILED_TOO_OLD = "O"
     FAILED_NO_DESTINATION = "D"
     FAILED_CHOICES = (
         (FAILED_SUSPENDED, _("Workspace suspended")),
+        (FAILED_CONTACT, _("Contact is no longer active")),
         (FAILED_LOOPING, _("Looping detected")),  # mailroom checks for this
         (FAILED_ERROR_LIMIT, _("Retry limit reached")),  # courier tried to send but it errored too many times
         (FAILED_TOO_OLD, _("Too old to send")),  # was queued for too long, would be confusing to send now